# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bringing group travel plans to life",
        "title": "Toggo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/files/confirm": {
            "post": {
                "description": "Verifies the file exists in S3 and marks the upload as confirmed in DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Confirm upload",
                "operationId": "confirmUpload",
                "parameters": [
                    {
                        "description": "Confirm upload request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConfirmUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConfirmUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/files/upload": {
            "post": {
                "description": "Generates presigned PUT URLs for uploading files and creates pending records in DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Create upload URLs",
                "operationId": "createUploadURLs",
                "parameters": [
                    {
                        "description": "Upload URL request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UploadURLRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.UploadURLResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/files/{imageId}": {
            "get": {
                "description": "Retrieves presigned URLs for all sizes of an image",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get all sizes of a file",
                "operationId": "getFileAllSizes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID (UUID)",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetFileAllSizesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes all size variants of an image from S3 and the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete image",
                "operationId": "deleteImage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID (UUID)",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/files/{imageId}/{size}": {
            "get": {
                "description": "Retrieves a presigned URL for downloading a specific file size",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get file by ID and size",
                "operationId": "getFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID (UUID)",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image size (large, medium, small)",
                        "name": "size",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.APIError": {
            "type": "object",
            "properties": {
                "message": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.ConfirmUploadRequest": {
            "type": "object",
            "required": [
                "imageId"
            ],
            "properties": {
                "imageId": {
                    "type": "string"
                },
                "size": {
                    "description": "Optional: if nil, confirm all sizes",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ImageSize"
                        }
                    ]
                }
            }
        },
        "models.ConfirmUploadResponse": {
            "type": "object",
            "properties": {
                "confirmed": {
                    "type": "integer"
                },
                "imageId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.GetFileAllSizesResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetFileResponse"
                    }
                },
                "imageId": {
                    "type": "string"
                }
            }
        },
        "models.GetFileResponse": {
            "type": "object",
            "properties": {
                "contentType": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                },
                "size": {
                    "$ref": "#/definitions/models.ImageSize"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.ImageSize": {
            "type": "string",
            "enum": [
                "large",
                "medium",
                "small"
            ],
            "x-enum-varnames": [
                "ImageSizeLarge",
                "ImageSizeMedium",
                "ImageSizeSmall"
            ]
        },
...
        "models.SizedUploadURL": {
            "type": "object",
            "properties": {
                "size": {
                    "$ref": "#/definitions/models.ImageSize"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.UploadURLRequest": {
            "type": "object",
            "required": [
                "contentType",
                "fileKey",
                "sizes"
            ],
            "properties": {
                "contentType": {
                    "type": "string",
                    "minLength": 1
                },
                "fileKey": {
                    "type": "string",
                    "minLength": 1
                },
                "sizes": {
                    "type": "array",
                    "maxItems": 3,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.ImageSize"
                    }
                }
            }
        },
        "models.UploadURLResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "fileKey": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                },
                "uploadUrls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SizedUploadURL"
                    }
                }
            }
        },
...
    }
}